requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
faiss-cpu
tensorflow-cpu
Pillow
